use theme everywhere, add icon support for nodes
